PC-95 Merge mint page into home page
